requirements.txt
langchain-astradb
langchain>=0.1.0
langchain-community 
langchain-groq
datasets
pypdf
python-dotenv
pandas
flask
awscli
queuelib
typing-extensions
-e .